+++ packages/update-changelog/index.js
@@ -1,6 +1,8 @@
 const fs = require('fs')
 const path = require('path')
 
+renameFolderUnreleased(process.env.npm_package_version)
+
 const filename = path.join(process.cwd(), 'CHANGELOG.md')
 if (!fs.existsSync(filename)) process.exit(0)
 
@@ -29,3 +31,22 @@ const updatedUnreleasedSection = unreleasedSubstring
 const newFileContent = fileContent.replace(unreleasedSubstring, updatedUnreleasedSection)
 
 fs.writeFileSync(filename, newFileContent)
+
+/**
+ * Rename folder _migrate/unreleased to _migrate/[normalized version]
+ *
+ * @param {string} version
+ */
+function renameFolderUnreleased (version) {
+  const normalizeVersion = require('./normalize-version')
+  const folderName = path.join(process.cwd(), '_migrate', 'unreleased')
+  const newFolder = path.join(process.cwd(), '_migrate', normalizeVersion(version))
+
+  if (fs.existsSync(folderName) && !fs.existsSync(newFolder)) {
+    try {
+      fs.renameSync(folderName, newFolder)
+    } catch (e) {
+      throw new Error(`Unable to rename ${folderName} to ${newFolder}: ${e}`)
+    }
+  }
+}