+++ packages/update-changelog/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Added
+- rename folders _migrate/unreleased in normalize version name (to match X.YY.ZZ version pattern)
+
 ## [1.0.2]
 ### Added
  - add changes types template sections just after [Unreleased] section. This prevents some types of merge conflicts in CHANGELOG's